requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zope.interface==6.3
